+++ cogdl/layers/__init__.py
@@ -8,14 +8,13 @@ from .se_layer import SELayer
 from .deepergcn_layer import GENConv, ResGNNLayer
 from .disengcn_layer import DisenGCNLayer
 from .gcnii_layer import GCNIILayer
-from .mlp_layer import MLPLayer
+from .mlp_layer import MLP
 from .saint_layer import SAINTLayer
 from .han_layer import HANLayer
 from .pprgo_layer import PPRGoLayer
 from .rgcn_layer import RGCNLayer
 from .sgc_layer import SGCLayer
 from .mixhop_layer import MixHopLayer
-from .mlp import MLP
 
 
 __all__ = [
@@ -32,7 +31,6 @@ __all__ = [
     "ResGNNLayer",
     "DisenGCNLayer",
     "GCNIILayer",
-    "MLPLayer",
     "SAINTLayer",
     "HANLayer",
     "PPRGoLayer",
